--- client/src/components/module/Account/ChangeAccountPassword/ChangeAccountPassword.module.css
@@ -1,11 +1,0 @@
-.container {
-    text-align: center;
-}
-
-.footer {
-    margin-top: 2em;
-}
-
-.secondary {
-    background-color: red;
-}
