--- client/src/components/module/Account/ChangeAccountPassword/index.js
@@ -1,1 +1,0 @@
-export { default as default } from "./ChangeAccountPassword";
