--- client/src/components/module/Account/EditEmployeeAccount/EditEmployeeAccount.module.css
@@ -1,7 +1,0 @@
-.container {
-    text-align: center;
-}
-
-.footer {
-    margin-top: 2em;
-}
