--- client/src/components/module/Account/EditEmployeeAccount/index.js
@@ -1,1 +1,0 @@
-export { default as default } from "./EditEmployeeAccount"
