+++ client/src/components/module/Account/index.js
@@ -1,11 +1,7 @@
 import AccountPreview from "./AccountPreview";
 import AccountButton from "./AccountButton";
-import EditEmployeeAccount from "./EditEmployeeAccount";
-import ChangeAccountPassword from "./ChangeAccountPassword";
 
 export default {
     AccountPreview,
     AccountButton,
-    EditEmployeeAccount,
-    ChangeAccountPassword
 };
